Moved Scheduler and avr devices into modules dir
