--- include/avr/device/stream/eeprom_stream.hpp
@@ -1,47 +1,0 @@
-/*
-    Aversive++                                                           
-    Copyright (C) 2014 Eirbot                                            
-                                                                         
-    This program is free software: you can redistribute it and/or modify 
-    it under the terms of the GNU General Public License as published by 
-    the Free Software Foundation, either version 3 of the License, or    
-    (at your option) any later version.                                  
-                                                                         
-    This program is distributed in the hope that it will be useful,      
-    but WITHOUT ANY WARRANTY; without even the implied warranty of       
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the        
-    GNU General Public License for more details.                         
-                                                                         
-    You should have received a copy of the GNU General Public License    
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#ifndef AVR_EEPROM_STREAM_HPP
-#define AVR_EEPROM_STREAM_HPP
-
-#include "../../../common/device/stream/eeprom_stream.hpp"
-
-#include <hardware/eeprom.hpp>
-
-inline EepromStream::EepromStream(const char* name) : Device(name) {
-}
-
-inline char EepromStream::getValue(void) {
-  if(_mini_buffer_used) {
-    _mini_buffer_used = false;
-    return _mini_buffer;
-  }
-  else {
-    char ret = Eeprom::read(cursor());
-    seek(cursor()+1);
-    return ret;
-  }
-}
-
-inline void EepromStream::setValue(char val) {
-  Eeprom::write(cursor(), val);
-  seek(cursor()+1);
-}
-
-
-#endif//AVR_EEPROM_STREAM_HPP
