--- include/avr/device/stream/uart_stream.hpp
@@ -1,49 +1,0 @@
-/*
-    Aversive++                                                           
-    Copyright (C) 2014 Eirbot                                            
-                                                                         
-    This program is free software: you can redistribute it and/or modify 
-    it under the terms of the GNU General Public License as published by 
-    the Free Software Foundation, either version 3 of the License, or    
-    (at your option) any later version.                                  
-                                                                         
-    This program is distributed in the hope that it will be useful,      
-    but WITHOUT ANY WARRANTY; without even the implied warranty of       
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the        
-    GNU General Public License for more details.                         
-                                                                         
-    You should have received a copy of the GNU General Public License    
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#ifndef AVR_UART_STREAM_HPP
-#define AVR_UART_STREAM_HPP
-
-#include "../../../common/device/stream/uart_stream.hpp"
-
-#include <hardware/uart.hpp>
-
-template<u8 _CHANNEL>
-UartStream<_CHANNEL>::UartStream(const char* name) : Device(name) {
-  Uart<_CHANNEL>::init();
-}
-
-template<u8 _CHANNEL>
-void UartStream<_CHANNEL>::setValue(char c) {
-  Uart<_CHANNEL>::send(c);
-}
-
-template<u8 _CHANNEL>
-char UartStream<_CHANNEL>::getValue(void) {
-  if(_mini_buffer_used) {
-    _mini_buffer_used = false;
-    return _mini_buffer;
-  }
-  else {
-    char ret;
-    Uart<_CHANNEL>::recv(ret);
-    return ret;
-  }
-}
-
-#endif//AVR_UART_STREAM_HPP
