+++ modules/avr/include/device/stream/eeprom_stream.hpp
@@ -39,4 +39,30 @@ public:
   void setValue(char c);
 };
 
+/////////////////////////////////////////////////////////////////////////////////
+
+
+#include <hardware/eeprom.hpp>
+
+inline EepromStream::EepromStream(const char* name) : Device(name) {
+}
+
+inline char EepromStream::getValue(void) {
+  if(_mini_buffer_used) {
+    _mini_buffer_used = false;
+    return _mini_buffer;
+  }
+  else {
+    char ret = Eeprom::read(cursor());
+    seek(cursor()+1);
+    return ret;
+  }
+}
+
+inline void EepromStream::setValue(char val) {
+  Eeprom::write(cursor(), val);
+  seek(cursor()+1);
+}
+
+
 #endif//EEPROM_STREAM_HPP
