+++ modules/avr/include/device/stream/uart_stream.hpp
@@ -45,4 +45,33 @@ public:
   void setValue(char c);
 };
 
+/////////////////////////////////////////////////////////////
+
+
+#include <hardware/uart.hpp>
+
+template<u8 _CHANNEL>
+UartStream<_CHANNEL>::UartStream(const char* name) : Device(name) {
+  Uart<_CHANNEL>::init();
+}
+
+template<u8 _CHANNEL>
+void UartStream<_CHANNEL>::setValue(char c) {
+  Uart<_CHANNEL>::send(c);
+}
+
+template<u8 _CHANNEL>
+char UartStream<_CHANNEL>::getValue(void) {
+  if(_mini_buffer_used) {
+    _mini_buffer_used = false;
+    return _mini_buffer;
+  }
+  else {
+    char ret;
+    Uart<_CHANNEL>::recv(ret);
+    return ret;
+  }
+}
+
+
 #endif//UART_STREAM_HPP
